requirements.txt

streamlit==1.18.1
opencv-python-headless==4.9.0.80
fer==22.5.1
numpy==1.26.4
Pillow
